backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic[email]==2.10.4
python-dotenv==1.0.1
python-multipart==0.0.12
